refactor testapp + fn in app.go
