+++ testing/testapp/wasm_1_simple.go
