+++ testing/testapp/wasm_2_return_in_prepare.go
