+++ testing/testapp/wasm_3_do_nothing.go
