+++ testing/testapp/wasm_9_set_data_several_times.go
