+++ testing/testapp/wasm_util.go
