+++ x/oracle/abci_test.go
@@ -10,7 +10,7 @@ package oracle_test
 // 	"github.com/stretchr/testify/require"
 // 	abci "github.com/tendermint/tendermint/abci/types"
 
-// 	"github.com/bandprotocol/chain/x/oracle/testapp"
+// 	"github.com/bandprotocol/chain/testing/testapp"
 // 	"github.com/bandprotocol/chain/x/oracle/types"
 // )
